requirements.txt
requests
pytest
pytest-html
python-dotenv
